+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add `version` constant to Application Class
 * Simplify `Application::invalidSessionRedirect`
 * Add Abstract Class `Controller`
+* Trait: `CallMethod`
 
 ## [v0.0.02](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.01...v0.0.02) - 8/26/2024